A workaround for org `eval-after-load` issue
org-mode passes string object as the first argument to
`eval-after-load`.  For example, check [1][]:

    (eval-after-load "flyspell"
      '(add-to-list 'flyspell-delayed-commands 'org-self-insert-command))

This form will be run after any file whose name is "flyspell" loaded.
If I have an init file called "flyspell.el", when this file is loaded,
Emacs will run the `add-to-list` form.  If `flyspell-delayed-commands`
has not been defined during the form execution, Emacs will raise an
error.

This workaround try to rename the conflict files to avoid the problem.

[1]: http://orgmode.org/cgit.cgi/org-mode.git/tree/lisp/org.el?id=1045e9e9c0e6438f5ee9dc4f0e5c720a8b670cdd#n25209