Force to not install some emacs packages
Some packages are built-in.  Some are sub-packages of others.

Set `:ensure nil` to force to not install them.  See [1].

In general: we need to make `req-package` install packages, so we have to set
`use-package-always-ensure` to `t`, so we have to declare `:ensure nil` for
these packages.

To learn the reason why we have to set `use-package-always-ensure`, See [2].

[1]: https://github.com/jwiegley/use-package/issues/388
[2]: https://github.com/edvorg/req-package/issues/51#issuecomment-362154387